zsh 11/28 done

+++ app/src/main/java/zzbcar/cckj/com/nzzb/bean/OrderBean.java
@@ -300,7 +300,7 @@ public class OrderBean implements Serializable {
             return trafficDepositMoney;
         }
 
-        public void setTrafficDepositMoney(int trafficDepositMoney) {
+        public void setTrafficDepositMoney(double trafficDepositMoney) {
             this.trafficDepositMoney = trafficDepositMoney;
         }
